:internal: fix delete key bug for text items

+++ egCAS/src/view/egcasscene.cpp
@@ -455,24 +455,28 @@ void EgCasScene::keyPressEvent(QKeyEvent* keyEvent)
         int key = keyEvent->key();
         if (key == Qt::Key_Delete || key == Qt::Key_Backspace) {
                 QList<QGraphicsItem*> list = selectedItems();
-                QGraphicsItem *item;
-                foreach (item, list) {
-                        if (    item->type() == static_cast<int>(EgcGraphicsItemType::EgcFormulaItemType)
-                             || item->type() == static_cast<int>(EgcGraphicsItemType::EgcPixmapItemType)
-                             || item->type() == static_cast<int>(EgcGraphicsItemType::EgcTextItemType)) {
-                                EgcFormulaItem *formula = dynamic_cast<EgcFormulaItem*>(item);
-                                if (formula) {  //formula
-                                        if (formula->aboutToBeDeleted()) {
+                if (list.isEmpty()) {
+                        QGraphicsScene::keyPressEvent(keyEvent);
+                } else {
+                        QGraphicsItem *item;
+                        foreach (item, list) {
+                                if (    item->type() == static_cast<int>(EgcGraphicsItemType::EgcFormulaItemType)
+                                                || item->type() == static_cast<int>(EgcGraphicsItemType::EgcPixmapItemType)
+                                                || item->type() == static_cast<int>(EgcGraphicsItemType::EgcTextItemType)) {
+                                        EgcFormulaItem *formula = dynamic_cast<EgcFormulaItem*>(item);
+                                        if (formula) {  //formula
+                                                if (formula->aboutToBeDeleted()) {
+                                                        deleteItem(item);
+                                                        m_document.itemDeleted(item);
+                                                        keyEvent->accept();
+                                                } else {  //call standard version that triggers the elements key event routine
+                                                        QGraphicsScene::keyPressEvent(keyEvent);
+                                                }
+                                        } else {  // pixmap or text
                                                 deleteItem(item);
                                                 m_document.itemDeleted(item);
                                                 keyEvent->accept();
-                                        } else {  //call standard version that triggers the elements key event routine
-                                                QGraphicsScene::keyPressEvent(keyEvent);
                                         }
-                                } else {  // pixmap or text
-                                        deleteItem(item);
-                                        m_document.itemDeleted(item);
-                                        keyEvent->accept();
                                 }
                         }
                 }
